- Added scheduler - Auto delnode when mem utilization in lower than threshhold

+++ src/main/java/com/blibli/caas/DTO/NodeStats.java
@@ -15,8 +15,8 @@ public class NodeStats implements Serializable {
   private String host;
   private String port;
   private String nodeId;
-  private long usedMemory;
-  private long totalMemory;
+  private double usedMemory;
+  private double totalMemory;
   private double usedCPU;
   private double totalCPU;
 }
